+++ apps/user/src/app/form/form.provider.ts
@@ -1,4 +1,3 @@
-import { FormStep } from '@/types/form';
 import { useRecoilState, atom } from 'recoil';
 
 const formDataAtomState = atom({
@@ -13,19 +12,9 @@ const formDataAtomState = atom({
     },
 });
 
-const formStepAtomState = atom<FormStep>({
-    key: 'form-step',
-    default: '지원자 정보',
-});
 
 export const useFormProvider = () => {
     const [form, setForm] = useRecoilState(formDataAtomState);
 
     return { form, setForm };
-};
-
-export const useFormStepProvider = () => {
-    const [formStep, setFormStep] = useRecoilState(formStepAtomState);
-
-    return { formStep, setFormStep };
 };
