+++ apps/user/src/app/form/성적입력/성적입력.hooks.ts
@@ -1,4 +1,5 @@
-import { useFormProvider, useFormStepProvider } from '../form.provider';
+import { useFormProvider } from '../form.provider';
+import { useFormStepProvider } from '@/provider/useFormStepProvider';
 import {
     useStudentSubjectListProvider,
     useAttendanceInfoProvider,
