+++ apps/user/src/app/form/자기소개서/자기소개서.hooks.ts
@@ -1,4 +1,5 @@
-import { useFormProvider, useFormStepProvider } from '../form.provider';
+import { useFormProvider } from '../form.provider';
+import { useFormStepProvider } from '@/provider/useFormStepProvider';
 import { useDebounceInput } from '@maru/hooks';
 
 export const useFormSubmitAction = (
