+++ apps/user/src/app/form/지원자정보/지원자정보.hooks.ts
@@ -1,4 +1,5 @@
-import { useFormProvider, useFormStepProvider } from '../form.provider';
+import { useFormProvider } from '../form.provider';
+import { useFormStepProvider } from '@/provider/useFormStepProvider';
 import formatDate, { Date } from '@/utils/formatDate';
 import { ChangeEventHandler, useEffect, useState } from 'react';
 import { useUserInfoProvider } from './지원자정보.provider';
