+++ apps/user/src/app/form/출신학교및학력/출신학교및학력.hooks.ts
@@ -1,4 +1,5 @@
-import { useFormProvider, useFormStepProvider } from '../form.provider';
+import { useFormProvider } from '../form.provider';
+import { useFormStepProvider } from '@/provider/useFormStepProvider';
 import { ChangeEventHandler } from 'react';
 import { useEducationInfoProvider } from './출신학교및학력.provider';
