+++ apps/user/src/provider/useFormStepProvider.ts
@@ -1,0 +1,13 @@
+import { FormStep } from '@/types/form';
+import { useRecoilState, atom } from 'recoil';
+
+const formStepAtomState = atom<FormStep>({
+    key: 'form-step',
+    default: '지원자 정보',
+});
+
+export const useFormStepProvider = () => {
+    const [formStep, setFormStep] = useRecoilState(formStepAtomState);
+
+    return { formStep, setFormStep };
+};
